Finally adding relative path for the Fable World project and cleaned the SETUP.txt file
The include and library paths are essential for the project, but the mystery was the additional library directories and the input libraries. For lua, the additional library directory has to be specified and lua5.1.lib has to be added as an input. For d3d9.lib;d3dx9.lib;dinput8.lib;dxerr.lib no additional library directory was needed, which was very suspicious, because on the other projects $(DXSDK_DIR)lib\x86 has to be specified again.
legacy_stdio_definitions.lib is needed for __vsnprintf and it is taken from somewhere from the internals of msvc (C:\Program Files (x86)\Microsoft Visual Studio\2019\Community\VC\Tools\MSVC\14.29.30037\lib\x86)