Removing more singletons
* Moved the GameObjectManager, DirectInput and TextManager under the
application. They are accessed from almost everywhere in the code base,
so I prefer they to be accessed through the application, instead of
being singletons. In the future if they can be localized in single
module they will be removed from Application and passed as parameter or
stored inside the classes as shared pointers, but for now this is not
the case

* moved GetStringWidth and GetStringHeight from Misc to TextManager

* removed leftover externs