Now using shared pointers instead of raw pointers for the game objects
* I could have used unique pointers for sure, but I like the agility
shared pointers give and also I have use case for shared ownership of a
game object - the main hero is stored in the game object manager, but
also in the Game scene, where it is used extensively.
* assignment and copy constructs of the game objects are now forbidden.
If a game object has to be made, it can be constructed via lua scripts
or SpawnClone
* removed the dreaded mainHero extern in LuaFunctions. It turned out to
be harder than expected, as I essentially want to access the Game scene
from the lua, while constructing the Game scene. In this sense better
approach turned out to be to check if a skinned model is a main hero,
while iterating them in the Game constructor. If a second main hero is
set by mistake the game exits with a message box.