The main hero is now automatically picked and the camera is adjusted to show the whole scene
* I still don't know how to imply that the game is currently with RTS
style and the player has to click on the terrain with the right mouse
button and go to Cho to kill it. With the previous RPG style it was more
obvious, as you pass by Ezreal, you get the quest for killing Cho and
then you encounter Cho. Probably moving to this RTS style was a mistake,
though I needed a showcase for the pathfinding. After the water
implementation, I can revert it back to the RPG style and will use the
pathfinding for the AI, will see.