Short refactoring of Terrain
* It all started with the invocation of GenerateVertexBuffer in
BuildSubGridMesh, which turned out not to be used.
* Moved the generation of index data in the subgrid, as previously it
was calculated twice
* The compute of the normals is now in the subgrid
* Added D3DXMESH_32BIT to the creation of the mesh in the subgrid. This
allows me to change the type of pIndices to DWORD, instead of WORD.
Using WORD can lead to very subtle and cryptic bugs.
* A question remains whether the vertex buffer can be created directly
in the subgrid, instead of having to create the whole terrain mesh,
which later to be discarded at the end of the function, in favor of the
subgrids. When I started the refactor I was left with the impression
that in the memory I have 2 terrains - one whole terrain and terrain
composed of subgrids. This is true when the terrain is constructed, but
the whole terrain is getting destroyed afterwards.