Adding the long-awaited water to the scene
* It uses two normal maps, which gives the illusion of moving waves,
though the surface is flat mesh.
* The generation of the flat mesh uses the same technique as with the
terrain - vertex and index buffer has to be generated again and water
has to be centered in the scene. I considered to move this commonality
in separate place, but the terrain goes further by dividing the big flat
mesh on submeshes for frustum culling, so I may add additional
complexity to already complex classes.
* More realistic water will be added next week via usage of displacement
maps