Adding more realistic water using displacement map and normal map
* it is based on the Frank D. Luna's book and on top of normal map,
displacement map is used so that the height of the water's vertices can
be changed in the vertex shader, based on the displacement maps, which
simulates waves. The displacement maps are sampled via tex2Dlod in the
vertex shader, which requires the shader model to be 3.0 instead of 2.0.
* more modern approaches for water can be seen here:
https://80.lv/articles/river-editor-water-simulation-in-real-time/, the
ideas being for wave particles, solving shallow-water differential
equations, Lattice Boltzmann methods.
* moving vertex declarations from Misc to Application, because it makes
more sense to be there
* fixing the shader's properties in Visual Studio. Turns out that the
shader type has to be specified, as well as the shader model and the
entry point name (not just the entry point name and the shader model).
It is optional for the shaders to be added in the project, but I like it
that way.